UserManagement/src/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-jose
passlib[bcrypt]
pydantic[email]
python-multipart